extract database quiz and results queries to seperate files, add loading animation to playpage, update play state
extract database quiz and resuluts queries to seperate files, add loading animation to playpage, update play state

edit comment

rm redundant condition

rm redundant condition
