+++ src/screens/PlayPage.js
@@ -1,4 +1,4 @@
-import { useEffect, useReducer } from 'react';
+import { useEffect, useReducer, useState } from 'react';
 import {
 	View,
 	Text,
@@ -8,47 +8,24 @@ import {
 } from 'react-native';
 import { SafeAreaView } from 'react-native-safe-area-context';
 import { styles } from '../styles/screens/PlayStyle';
-import { db } from '../../firebaseConfig';
-import { query, getDocs, collection, where } from 'firebase/firestore';
+import { Option, GoBack, LoadingAnimation } from '../components/Index';
 import { QuizReducer, INITIAL_STATE } from '../utilities/QuizReducer';
-import { Option, GoBack } from '../components/Index';
+import { fetchQuiz } from '../utilities/fetchQuiz';
 
 const PlayPage = ({ route, navigation }) => {
 	// Contains all relevant information on the specified quiz, see: ../utilities/QuizReducer
 	const [state, dispatch] = useReducer(QuizReducer, INITIAL_STATE);
+	//const [isLoading, setIsLoading] = useState(true);
 
-	// Retrieve all questions related to the specified quiz then update state with information
+	// Retrieve all questions from specified quiz then set state with information on first render
 	useEffect(() => {
-		// Retrieve specified quiz as parameter from route
-		const { number, quiz } = route.params;
-		// Query all documents in questions collection containing a link to the specified quiz
-		const questionsQuery = query(
-			collection(db, 'questions'),
-			where('quizzes', 'array-contains', quiz)
+		const { quiz } = route.params;
+		fetchQuiz(quiz, dispatch).then(() =>
+			dispatch({ type: 'setisloading', payload: false })
 		);
-
-		const fetchData = async () => {
-			const querySnapshot = await getDocs(questionsQuery);
-			// Append each question retrieved from database to state
-			querySnapshot.forEach((doc) => {
-				dispatch({ type: 'setquestionsarray', payload: doc.data() });
-			});
-
-			// Update remaining state information responsible for rendering text and handling logic
-			dispatch({
-				type: 'setmulitple',
-				payload: {
-					quizLength: querySnapshot.docs.length,
-					questionText: querySnapshot.docs[0].data().question_text,
-					options: querySnapshot.docs[0].data().options,
-					correctOption: querySnapshot.docs[0].data().correct_answer,
-				},
-			});
-		};
-
-		fetchData().catch((error) => console.log(error));
 	}, []);
 
+	// Continue to next question on answer input
 	useEffect(() => {
 		try {
 			dispatch({
@@ -65,46 +42,50 @@ const PlayPage = ({ route, navigation }) => {
 	}, [state.index]);
 
 	return (
-		<ImageBackground
-			source={require('../assets/images/play_bg.png')}
-			style={{ flex: 1, width: null, alignSelf: 'stretch' }}
-		>
-			<SafeAreaView>
-				<GoBack nav={navigation} destination={'homepage'} />
-				{state.index < state.quizLength && (
-					<Text style={styles.IndexText}>
-						Spørsmål {state.index + 1} av {state.quizLength}
-					</Text>
-				)}
-				{state.index < state.quizLength ? (
-					<View>
-						<Text style={styles.QuestionText}>{state.questionText}</Text>
-						{state.options.map((option, idx) => (
-							<Option
-								value={option}
-								key={idx}
-								id={idx}
-								state={state}
-								dispatch={dispatch}
-							/>
-						))}
-					</View>
-				) : (
-					<View>
-						<Text style={styles.QuestionText}>
-							Du har fullførten quizen. Gå videre for å se resultatene dine.
-						</Text>
-						<TouchableOpacity
-							style={styles.resultsBtn}
-							onPress={() => navigation.navigate('resultspage')}
-						>
-							<Text style={styles.btnText}>Resultater</Text>
-						</TouchableOpacity>
-					</View>
-				)}
-			</SafeAreaView>
-			<StatusBar translucent backgroundColor="transparent" />
-		</ImageBackground>
+		<View style={styles.containerTo}>
+			{state.isLoading ? (
+				<LoadingAnimation />
+			) : (
+				<ImageBackground
+					source={require('../assets/images/play_bg.png')}
+					style={{ flex: 1, width: null, alignSelf: 'stretch' }}
+				>
+					<SafeAreaView>
+						<GoBack nav={navigation} destination={'homepage'} />
+						{state.index < state.quizLength ? (
+							<View>
+								<Text style={styles.IndexText}>
+									Spørsmål {state.index + 1} av {state.quizLength}
+								</Text>
+								<Text style={styles.QuestionText}>{state.questionText}</Text>
+								{state.options.map((option, idx) => (
+									<Option
+										value={option}
+										key={idx}
+										id={idx}
+										state={state}
+										dispatch={dispatch}
+									/>
+								))}
+							</View>
+						) : (
+							<View>
+								<Text style={styles.QuestionText}>
+									Du har fullført quizen. Gå videre for å se resultatene dine.
+								</Text>
+								<TouchableOpacity
+									style={styles.resultsBtn}
+									onPress={() => navigation.navigate('resultspage')}
+								>
+									<Text style={styles.btnText}>Resultater</Text>
+								</TouchableOpacity>
+							</View>
+						)}
+					</SafeAreaView>
+					<StatusBar translucent backgroundColor="transparent" />
+				</ImageBackground>
+			)}
+		</View>
 	);
 };
