+++ src/styles/screens/PlayStyle.js
@@ -1,12 +1,15 @@
 import { StyleSheet } from 'react-native';
 
 export const styles = StyleSheet.create({
+	containerTo: {
+		flex: 1,
+	},
 	IndexText: {
 		position: 'absolute',
 		color: '#000',
 		fontWeight: '400',
 		fontSize: 14,
-		bottom: 65,
+		bottom: 30,
 		paddingHorizontal: 20,
 	},
 	QuestionText: {
